requirements.txt
Django>=4
django-extensions
django-versionfield
django-crispy-forms
crispy-bootstrap5
django-debug-toolbar